Moved illustrations into a new sub-folder; update paths
